backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.43
alembic==1.16.5
psycopg[binary]==3.2.10
pydantic==2.11.7
pydantic-settings==2.10.1
python-multipart==0.0.20
geoalchemy2==0.18.0
geopandas==1.1.1
shapely==2.1.1
h3==4.3.1
redis==6.4.0
celery==5.5.3
slowapi==0.1.9
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
httpx==0.28.1
prometheus-fastapi-instrumentator==7.1.0
orjson==3.11.3
